Mean free path Python wrapper
